requirements.txt
pandas~=2.0.3
scikit-learn~=1.3.0

requests~=2.31.0
python-dotenv~=1.0.0